* Adding expression evaluator * Integrated with UI

+++ src/main/java/core/operations/Operation.java
@@ -1,5 +1,7 @@
 package core.operations;
 
 public interface Operation {
-    long exec(long n1, long n2);
+    double exec(double n1, double n2);
+
+
 }
